+++ source/production/F0.Templates/templates/Roslyn-Generator/Roslyn.Generator.Example/Greeter.cs
@@ -1,0 +1,10 @@
+using Roslyn.Generated;
+
+namespace Roslyn.Generator.Example
+{
+	internal static partial class Greeter
+	{
+		[HelloWorld]
+		public static partial string GetHelloWorld();
+	}
+}