+++ source/production/F0.Templates/templates/Roslyn-Generator/Roslyn.Generator.Example/Program.cs
@@ -1,0 +1,12 @@
+namespace Roslyn.Generator.Example
+{
+	internal static class Program
+	{
+		private static void Main(string[] args)
+		{
+			string greeting = Greeter.GetHelloWorld();
+
+			Console.WriteLine(greeting);
+		}
+	}
+}